registry: fix no size because of foward reference error
When trying to utilize the BasicRegistry aliases in another directory outside the ecs lib,
as a function param a "no size because of forward reference" error happens. Code example:

```d
import ecs;
void foo(Registry registry) {}
```

This was found when using the ecs lib as a dependency in another project. The project was
strutured the following way:

```
project
   |--- source
   |      |--- dub.json
   |      |--- ecs
   |            |--- (...)
   |--- examples
           |--- dub.json
           |--- <dir>
                  |--- dub.json (using ecs as a dependency)
                  |--- source
                         |--- (...)
```

This cause is unknown, however, a workaround was found when trying to find the core issue.
Solution: add the following BasicRegistry class template validation:

```d
__traits(compiles, SparseSet!(T, idBitAmount))
```

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>